// apps/api/src/models/types/public/Runs.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { Id } from './Monitors';
import type { default as RunStatus } from './RunStatus';

/** Identifier type for runs */
export type Id = string;

/** Represents the table public.runs */
export default interface Runs {
  id: Id;

  monitor_id: Id;

  status: RunStatus;

  duration_ms: number;

  result_details: unknown | null;

  created_at: Date;
}

/** Represents the initializer for the table public.runs */
export interface RunsInitializer {
  id: Id;

  monitor_id: Id;

  status: RunStatus;

  /** Default value: 0 */
  duration_ms?: number;

  result_details?: unknown | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;
}

/** Represents the mutator for the table public.runs */
export interface RunsMutator {
  id?: Id;

  monitor_id?: Id;

  status?: RunStatus;

  duration_ms?: number;

  result_details?: unknown | null;

  created_at?: Date;
}